requirements.txt
asgiref==3.8.1
colorama==0.4.6
Django==5.2.3
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
iniconfig==2.1.0
joblib==1.5.1
numpy==2.3.1
packaging==25.0
pluggy==1.6.0
psycopg2-binary==2.9.10
Pygments==2.19.2
PyJWT==2.9.0
pytest==8.4.1
pytest-django==4.11.1
scikit-learn==1.7.0
scipy==1.16.0
sqlparse==0.5.3
threadpoolctl==3.6.0
tzdata==2025.2
